updated mapping

+++ kmertools.scala
@@ -7,6 +7,21 @@ package hse
 
 import Stream._
 
+/////////////////////////////////////////////////////////////////////////////
+
+case class kmerType(seq: String, index: Int) {
+  /*************************************************************************
+   * Define a kmertype such that we keep the index together with the kmer  *
+   *   sequence. We have to redefine the equality and hashCode functions   *
+   *   such that this type works with the sets and hashtables              *
+   *************************************************************************/
+  override def equals(o: Any) = o match {
+    case that: kmerType => that.seq.equalsIgnoreCase(this.seq)
+    case _ => false
+  }
+  override def hashCode = seq.toUpperCase.hashCode
+}
+
 ///////////////////////////////////////////////////////////////////////////////
 
 object kmerTools {
@@ -22,18 +37,18 @@ object kmerTools {
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def kmer_at_i(seq: String, i: Int, upst_downst: Int): String = {
+  def kmer_at_i(seq: String, i: Int, upst_downst: Int): kmerType = {
     /*************************************************************************
      * Return a substring based at location i, with up and downstream        *
      *   limits                                                              *
      *************************************************************************/
-    seq.substring(i-upst_downst, i+upst_downst+1)
+    new kmerType(seq.substring(i-upst_downst, i+upst_downst+1), i)
   }
 
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def allKmersSet(seq: String, k: Int): Set[String] = {
+  def allKmersSet(seq: String, k: Int): Set[kmerType] = {
     /*************************************************************************
      * Return all the Kmers in a string seq, given a size of k, as a set     *
      *************************************************************************/
@@ -47,7 +62,7 @@ object kmerTools {
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def allKmersStream(seq: String, k: Int): Stream[String] = {
+  def allKmersStream(seq: String, k: Int): Stream[kmerType] = {
     /*************************************************************************
      * Return all the Kmers in a string seq, given a size of k, as a stream  *
      *************************************************************************/
@@ -55,7 +70,7 @@ object kmerTools {
     val upst_downst = ((k.toDouble - 1.0) / 2.0).toInt
     val limit       = seq.length - upst_downst - 1
 
-    def allKmersStreamHelper(seq: String, i: Int): Stream[String] = {
+    def allKmersStreamHelper(seq: String, i: Int): Stream[kmerType] = {
       if (i > limit){
        return empty
       }
@@ -68,24 +83,30 @@ object kmerTools {
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def uniqueKmers(kmers: List[Set[String]]): List[Set[String]] = {
+  def uniqueKmers(kmers: List[Set[kmerType]]): List[Set[kmerType]] = {
     /**************************************************************************
      * Given a list of outputs from allKmersSet, return only those which are  *
      *   unique to one of them.                                               *
+     * Special care if there is only one sample, then there is no need        *
      **************************************************************************/
-     val kmer_index : List[(Set[String],Int)]   = kmers.zipWithIndex;
-     val remove_index = (L: List[(Set[String],Int)], index: Int) =>(L.patch(index, Nil, 1))
+     if (kmers.length > 1){
+       val kmer_index : List[(Set[kmerType],Int)]   = kmers.zipWithIndex;
+       val remove_index = (L: List[(Set[kmerType],Int)], index: Int) =>(L.patch(index, Nil, 1))
      
-     def setUnion(L: List[Set[String]]): Set[String] = {
-       var U = L(0);
-       L.fold(U){ (a,b) => a.union(b) }
+       def setUnion(L: List[Set[kmerType]]): Set[kmerType] = {
+         var U = L(0);
+         L.fold(U){ (a,b) => a.union(b) }
+       }
+       kmer_index.map(elem => elem._1.diff(setUnion(remove_index(kmer_index, elem._2).map(vi => vi._1))))
+     } else {
+       // We don't need to do the above, because it is a lonely species
+       kmers
      }
-     kmer_index.map(elem => elem._1.diff(setUnion(remove_index(kmer_index, elem._2).map(vi => vi._1))))
   }
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def orthologUniqueKmers(orthologs: List[Fasta.Entry], k: Int) : (List[String], List[Set[String]]) = {
+  def orthologUniqueKmers(orthologs: List[Fasta.Entry], k: Int) : (List[String], List[Set[kmerType]]) = {
     /*************************************************************************
      * Given a list of orthologs, return sets of unique kmers for each       *
      *   ortholog.                                                           *
@@ -101,9 +122,9 @@ object kmerTools {
   /////////////////////////////////////////////////////////////////////////////
 
 
-  def genomeUniqueKmers(kmers: Set[String], gKmers: genomeKmers): Set[String] = {
-    def genomeUniqueKmerHelper(kmer: String): List[String] = {
-      if(gKmers.countOccurences(kmer) > 1){
+  def genomeUniqueKmers(kmers: Set[kmerType], gKmers: genomeKmers): Set[kmerType] = {
+    def genomeUniqueKmerHelper(kmer: kmerType): List[kmerType] = {
+      if(gKmers.countOccurences(kmer.seq) > 1){
         Nil
       }else{
         List(kmer)
@@ -114,21 +135,31 @@ object kmerTools {
 
   /////////////////////////////////////////////////////////////////////////////
 
-  def orthologGenomeUniqueKmers(oKmers: List[(List[String], List[Set[String]])], gKmers: genomeKmers): List[(List[String], List[Set[String]])] = {
+  def orthologGenomeUniqueKmers(oKmers: List[(List[String], List[Set[kmerType]])], gKmers: genomeKmers): List[(List[String], List[Set[kmerType]])] = {
      oKmers.map(x => (x._1, x._2.map( y => genomeUniqueKmers(y, gKmers))))
   }
 
   /////////////////////////////////////////////////////////////////////////////
+
+  //val oKmers = Set(new kmerType("ABCDE", 2), new kmerType("CDEFG", 4), new kmerType("WXYZA", 20))
+  //val kmers = oKmers
+  //val kmersList = kmers.toList.sortWith(_.index < _.index)
+  //kmersList.foldLeft(List.empty[kmerType]){ (a,b) => {if (a.length == 0 || (a.last.index + k) < b.index){ a :+ b }else{ a }}}.toSet
+
+  def nonRedundantKmers(oKmers: List[(List[String], List[Set[kmerType]])], k: Int): List[(List[String], List[Set[kmerType]])] = {
+    def nonRedundantKmersHelper(kmers: Set[kmerType]): Set[kmerType] = {
+      val kmersList = kmers.toList.sortWith(_.index < _.index)
+      kmersList.foldLeft(List.empty[kmerType]){ (a,b) => {if (a.length == 0 || (a.last.index + k) < b.index){ a :+ b }else{ a }}}.toSet
+    }
+
+    oKmers.map(x => (x._1, x._2.map(y => nonRedundantKmersHelper(y))))
+  }
 
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
-
-// Just so I can refer to them without all the stuff
-import collection.mutable.{HashMap  => mutHashMap}
-import collection.immutable.{HashMap => imHashMap}
 
 //class genomeKmers(fastas: List[Fasta.Entry], k: Int) {
 ///*****************************************************************************
@@ -192,6 +223,11 @@ import collection.immutable.{HashMap => imHashMap}
 //
 //}
 
+// Just so I can refer to them without all the stuff
+import collection.mutable.{HashMap  => mutHashMap}
+import collection.immutable.{HashMap => imHashMap}
+
+
 class genomeKmers(fastas: List[Fasta.Entry], k: Int) {
 /*****************************************************************************
  * A class to handle the kmer HashMap for genome sequences                   *
@@ -214,16 +250,16 @@ class genomeKmers(fastas: List[Fasta.Entry], k: Int) {
     
     for(fasta <- fastas){
       val kmerIterator = kmerTools.allKmersStream(fasta.sequence, k).iterator
-      var kmer  = new String("  ")
+      var kmer  = new kmerType("  ", 0)
       var count = 0
       
       while(kmerIterator.hasNext){
         kmer  = kmerIterator.next()
         count = 0
-        if (kmerCounts.contains(kmer)){
-          count = kmerCounts(kmer)
+        if (kmerCounts.contains(kmer.seq)){
+          count = kmerCounts(kmer.seq)
         }
-        kmerCounts += Tuple2(kmer,count + 1);
+        kmerCounts += Tuple2(kmer.seq,count + 1);
       }
     }
     
@@ -244,6 +280,6 @@ class genomeKmers(fastas: List[Fasta.Entry], k: Int) {
     }
   }
 
-
+}
 ///////////////////////////////////////////////////////////////////////////////
